requirements.txt
# Core Computer Vision Libraries
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0,<2.0.0

# OpenAI for Text-to-Speech
openai>=1.0.0

# Audio Processing
pygame>=2.5.0  # For playing audio
pydub>=0.25.0  # Audio manipulation (optional)

# Machine Learning (for gesture classification)
scikit-learn>=1.3.0
joblib>=1.3.0  # For saving/loading models

# Data Processing
pandas>=2.0.0  # Optional, for data analysis

# Configuration Management
python-dotenv>=1.0.0  # For managing environment variables

# GUI (optional, can use OpenCV or upgrade to more advanced)
# tkinter is built-in with Python

# Development Tools
pytest>=7.4.0  # For testing
black>=23.0.0  # Code formatter
flake8>=6.0.0  # Linter

# Progress Bar (nice to have)
tqdm>=4.65.0

# Logging
loguru>=0.7.0
